REFACTOR: Modal & ModalSearchBar 폴더 변경
